authentication

+++ ApiDotNet.Api/Controllers/PersonController.cs
@@ -1,5 +1,6 @@
 ﻿using ApiDotNet.Application.DTOs;
 using ApiDotNet.Application.Services.Interfaces;
+using ApiDotNet.Domain.Authentication;
 using ApiDotNet.Domain.FiltersDb;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -9,19 +10,27 @@ namespace ApiDotNet.Api.Controllers
     [Authorize]
     [Route("api/[controller]")]
     [ApiController]
-    public class PersonController : ControllerBase
+    public class PersonController : BaseController
     {
         //importar serviços
         private readonly IPersonService _personService;
+        private readonly ICurrentUser _currentUser;
+        private List<string> _permissionNeeded = new List<string>(){"Admin"};
+        private readonly List<string> _permissionUser;
 
-        public PersonController(IPersonService personService)
+        public PersonController(IPersonService personService, ICurrentUser currentUser)
         {
             _personService = personService;
+            _currentUser = currentUser;
+            _permissionUser = _currentUser?.Permissions?.Split(",")?.ToList() ?? new List<string>();
         }
 
         [HttpPost]
-        public async Task<ActionResult> PostAsync([FromBody] PersonDTO personDTO)
+        public async Task<IActionResult> PostAsync([FromBody] PersonDTO personDTO)
         {
+            _permissionNeeded.Add("CadastraPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
             var result = await _personService.CreateAsync(personDTO);
             if (result.IsSuccess)
                 return Ok(result);
@@ -30,8 +39,11 @@ namespace ApiDotNet.Api.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult> GetAsync()
+        public async Task<IActionResult> GetAsync()
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
             var result = await _personService.GetAsync();
             if (result.IsSuccess)
                 return Ok(result);
@@ -41,8 +53,11 @@ namespace ApiDotNet.Api.Controllers
 
         [HttpGet]
         [Route("{id}")]
-        public async Task<ActionResult> GetByIdAsync(int id)
+        public async Task<IActionResult> GetByIdAsync(int id)
         {
+            _permissionNeeded.Add("BuscaPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
             var result = await _personService.GetByIdAsync(id);
             if (result.IsSuccess)
                 return Ok(result);
@@ -51,8 +66,11 @@ namespace ApiDotNet.Api.Controllers
         }
 
         [HttpPut]
-        public async Task<ActionResult> UpdateAsync([FromBody] PersonDTO personDTO)
+        public async Task<IActionResult> UpdateAsync([FromBody] PersonDTO personDTO)
         {
+            _permissionNeeded.Add("EditaPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
             var result = await _personService.UpdateAsync(personDTO);
             if (result.IsSuccess)
                 return Ok(result);
@@ -62,8 +80,11 @@ namespace ApiDotNet.Api.Controllers
 
         [HttpDelete]
         [Route("{id}")]
-        public async Task<ActionResult> DeleteAsync(int id)
+        public async Task<IActionResult> DeleteAsync(int id)
         {
+            _permissionNeeded.Add("DeletaPessoa");
+            if (!ValidPermission(_permissionUser, _permissionNeeded))
+                return Forbidden();
             var result = await _personService.DeleteAsync(id);
             if (result.IsSuccess)
                 return Ok(result);
